+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @napplet/core
 
+## 0.5.0
+
+### Minor Changes
+
+- Align media APIs with ownership-aware NAP-MEDIA sessions.
+
 ## 0.4.0
 
 ### Minor Changes